Revert Braxton code refactor
