+++ app/sampl/base_settings.py
@@ -1,6 +1,9 @@
 """
 Django settings for SAMPL project.
 """
+import os
+import subprocess
+
 # Custom settings for SAMPL
 VISUALIZE_DASK_GRAPH = False
 DASK_SCHEDULER_URL = "localhost:8786"
@@ -104,7 +107,46 @@ REMOTE_SCHEDULER = False
 # container engine is either "docker" or "singularity"
 CONTAINER_ENGINE = "docker"
 
+ECR_BASE_URL = os.environ["ECR_BASE_URL"]
+ECR_SAMPLLEAGUE_URL = os.environ["ECR_SAMPLLEAGUE_URL"]
+
+
+def run_aws_login(container_engine):
+    if container_engine == "docker":
+        subprocess.run(
+            [
+                f"aws ecr get-login-password --region us-east-2 --profile sampl_pull | docker login --username AWS --password-stdin {ECR_BASE_URL}"
+            ],
+            shell=True,
+            check=True,
+        )
+    elif container_engine == "singularity":
+        os.environ["SINGULARITY_DOCKER_USERNAME"] = "AWS"
+        aws_call = subprocess.run(
+            [
+                "aws",
+                "ecr",
+                "get-login-password",
+                "--region",
+                "us-east-2",
+                "--profile",
+                "sampl_pull",
+            ],
+            capture_output=True,
+            text=True,
+            check=True,
+        )
+        os.environ["SINGULARITY_DOCKER_PASSWORD"] = aws_call.stdout
+    else:
+        raise Exception("Container Engine not yet implemented")
+
+
+def run_aws_logout(container_engine):
+    if container_engine == "singularity":
+        os.environ["SINGULARITY_DOCKER_USERNAME"] = ""
+        os.environ["SINGULARITY_DOCKER_PASSWORD"] = ""
+
 
 LOGIN_TO_AWS = False
-AWS_LOGIN_FUNCTION = None
-AWS_LOGOUT_FUNCTION = None
+AWS_LOGIN_FUNCTION = run_aws_login
+AWS_LOGOUT_FUNCTION = run_aws_logout
